+++ lib/neo4j_dart.dart
@@ -5,7 +5,9 @@ library neo4j_dart;
 
 import 'dart:async';
 import 'dart:convert';
+import 'package:crypto/crypto.dart';
 import 'package:http/http.dart' as http;
 
 part 'src/db.dart';
 part 'src/statement.dart';
+part 'src/transaction.dart';